+++ tests/ClassSchedule.UI.Shared.Tests/FakeViewModel.cs
@@ -1,7 +1,0 @@
-#pragma warning disable IDE0130
-namespace ClassSchedule.UI.Shared.ViewModels;
-
-/// <summary>
-/// 测试用视图模型
-/// </summary>
-public sealed class FakeViewModel;